Optimize render cycles of GraphNode and GraphEdge components.
